Renaming client-shared-server netwok components
