+++ src/main/client/player/player-manager.ts
@@ -2,7 +2,7 @@ import { Inject, Singleton } from 'typescript-ioc';
 import { PlayerComponent } from './player-component';
 import { GameStateComponent } from '../game-state/game-state-component';
 import { GameComponent } from '../game/game-component';
-import { NetworkComponent } from '../network/network-component';
+import { ClientNetworkComponent } from '../network/client-network-component';
 
 @Singleton
 export class PlayerManager {
@@ -10,7 +10,7 @@ export class PlayerManager {
       @Inject private readonly component: PlayerComponent,
       @Inject private readonly gameState: GameStateComponent,
       @Inject private readonly game: GameComponent,
-      @Inject private readonly network: NetworkComponent,
+      @Inject private readonly network: ClientNetworkComponent,
    ) {
       network.loginResponse$.subscribe((response) => {
          console.log('My id is ', response.id);
