+++ src/main/server/network/server-network-model.ts
@@ -1,4 +1,4 @@
-import { NetworkMessage } from '../../client/network/network-model';
+import { NetworkMessage } from '../../shared/network/shared-network-model';
 
 export interface ServerNetworkMessage extends NetworkMessage {
    user: string;
